Rename class "FEFUPascalCompiler" to "Compiler"
